% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_kinases}
\alias{human_kinases}
\title{human_kinases}
\format{tibble}
\usage{
human_kinases
}
\description{
A tibble containing  a catalog of human kinases.
}
\details{
Derived from an updated version of the kinases described here:

The Protein Kinase Complement of the Human Genome
G Manning, DB Whyte, R Martinez, T Hunter, S Sudarsanam (2002).
Science 298:1912-1934

source: http://kinase.com/human/kinome/tables/Kincat_Hsap.08.02.xls
}
\keyword{datasets}
